backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
email-validator
bcrypt==4.0.1
websockets
httpx
